LayoutTests/platform/chromium/accessibility/post-notification-ActiveDescendantChanged.html

<html>
<head>
<script>
    function testNotification() {
        var radiogroup = document.getElementById("ul-for-ActiveDescendantChanged");

        radiogroup.focus();
        radiogroup.setAttribute("aria-activedescendant", "r1");
    }
</script>
</head>
<body onload="test()">
This tests that chromium correctly recieves the ActiveDescendantChanged notification.

<ul id="ul-for-ActiveDescendantChanged" aria-help="ul-for-ActiveDescendantChanged" tabindex="0" role="radiogroup">
    <li id="r1">li1</li>
</ul>

<script type="text/javascript" src="post-notification.js"></script>
</body>
</html>
