LayoutTests/platform/chromium/accessibility/post-notification-ActiveDescendantChanged-expected.txt

AccessibilityNotification - FocusedUIElementChanged - id:ul-for-ActiveDescendantChanged
AccessibilityNotification - FocusedUIElementChanged - id:r1
AccessibilityNotification - ActiveDescendantChanged - id:ul-for-ActiveDescendantChanged
AccessibilityNotification - LoadComplete
This tests that chromium correctly recieves the ActiveDescendantChanged notification.
li1
